Created a user-input system, began work on a class selection system.
Added a generic title.

Added a user-input system where a character can be named (Note: after hitting submit, there is no logic that saves the player's name or class selected, this is being worked on).

Created pixel art for the Thorian Knight class, and made a function that changes the color of the class border upon clicking on a class. This indicates that the player has a class currently selected.