Reduced amount of JS, CSS and HTML changes...
Removed a significant portion of Javascript code and added some css and html, in favor of simplicity to achieve the same result - user can now only choose one class, not 2 or all 3. This is achieved through the use of hidden radio buttons, which are then hidden through CSS.

Additionally, the placement of some elements have changed, and there is a new custom-made pixel background.